Add support for looking up external identifiers with slashes
The code inside BagsApi was getting fairly knarly, so I broke it into
a couple of Lookup traits.  This lets BagsApi focus on routing, which
turns out to be a bit fiddly to ensure we don't consume "/versions" as
part of the external identifier.

Includes tests that we can look up a bag with slashes in the identifier
that are both URL-encoded and not

i.e. /bags/alfa/bravo and /bags/alfa%2Fbravo
